pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "circuit-knitting-toolbox"
version = "0.2.0"
description = "A software prototype for a circuit knitting toolbox which connects user applications with runtime primitives"
readme = "README.md"
license = {file = "LICENSE.txt"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Physics",
]

requires-python = ">=3.8"

dependencies = [
    "qiskit-aer>=0.12.0",
    "qiskit-terra>=0.24.0",
    "qiskit-nature>=0.6.0",
    "qiskit-ibm-runtime>=0.9.2",
]

[project.optional-dependencies]
cplex = [
    # We use the same restrictions in both of the following lines, as there
    # is no reason for us to install docplex without cplex.
    "docplex>=2.23.222; python_version < '3.11' and platform_machine != 'arm64'",
    "cplex>=22.1.0.0; python_version < '3.11' and platform_machine != 'arm64'",
]
gurobi = [
    "gurobipy==11.0.0"
]
pyscf = [
    "pyscf>=2.0.1; sys_platform != 'win32'",
]
dev = [
    "circuit-knitting-toolbox[test,nbtest,lint]",
]
basetest = [
    "pytest>=6.2.5",
    "pytest-randomly>=1.2.0",
]
test = [
    "circuit-knitting-toolbox[basetest,pyscf]",
    "ddt>=1.4.4",
]
nbtest = [
    "circuit-knitting-toolbox[basetest]",
    "nbmake>=1.3.4"
]
style = [
    "autoflake==2.1.1",
    "black[jupyter]==23.3.0",
    "ruff>=0.0.246",
    "nbqa>=1.6.0",
]
lint = [
    "circuit-knitting-toolbox[style]",
    "pydocstyle==6.3.0",
    "mypy==1.3.0",
    "reno>=3.4.0",
    # pydocstyle prefers to parse our pyproject.toml, hence the following line
    "toml",
]
docs = [
    "Sphinx>=3.0.0",
    "sphinx-autodoc-typehints>=1.12.0",
    "sphinx-reredirects",
    "jupyter-sphinx>=0.3.2",
    "nbsphinx>=0.8.8",
    "sphinx-copybutton>=0.5.0",
    "reno>=3.4.0",
]
notebook-dependencies = [
    "circuit-knitting-toolbox[cplex,pyscf,gurobi]",
    "quantum-serverless>=0.0.7",
    "matplotlib",
    "ipywidgets",
    "pylatexenc",
]

[project.urls]
"Documentation" = "https://qiskit-extensions.github.io/circuit-knitting-toolbox/"
"Repository" = "https://github.com/Qiskit-Extensions/circuit-knitting-toolbox"

[tool.hatch.build.targets.wheel]
only-include = [
    "circuit_knitting",
    "circuit_knitting_toolbox",
]

[tool.autoflake]
remove-unused-variables = true
remove-all-unused-imports = true

[tool.coverage.run]
omit = [
    # deprecated import location(s)
    "circuit_knitting/cutting/wire_cutting.py",
]

[tool.pytest.ini_options]
testpaths = ["./circuit_knitting/", "./test/"]
filterwarnings = ["ignore:::.*qiskit.opflow*"]
addopts = "--doctest-modules -rs --durations=10"
